deployment/requirements.txt
pandas
joblib
pillow
sklearn
streamlit